test(ci): split plugin bump advisory coverage
Keep advisory-specific tests in a small file while preserving the warning and exit-code cases for unnecessary plugin manifest bumps.
